+++ www/go/base/model/Acl.php
@@ -118,12 +118,6 @@ class Acl extends \GO\Base\Db\ActiveRecord {
 		if(\GO::$ignoreAclPermissions && $userId===false)
 			return self::MANAGE_PERMISSION;
 
-		if( \GO::user() && \go\core\model\User::isAdminById($userId ? $userId : \GO::user()->id)) {
-
-			return $aclId == \go\core\model\Acl::getReadOnlyAclId() ? self::READ_PERMISSION : self::MANAGE_PERMISSION ;
-
-		}
-
 		if($userId===false){
 			if(\GO::user())
 				$userId=\GO::user()->id;