+++ www/go/core/Installer.php
@@ -114,6 +114,7 @@ class Installer {
 		go()->clearCache();
 		$cacheCls = get_class(go()->getCache());
 		go()->setCache(new None());
+		go()->disableEvents();
 
 
 		self::$isInstalling = true;
@@ -181,6 +182,8 @@ class Installer {
 		jmap\Entity::$trackChanges = true;
 		go()->getDbConnection()->exec("SET FOREIGN_KEY_CHECKS=1;");
 		self::$isInstalling = false;
+
+		go()->enableEvents();
 	}
 
 	/**
@@ -472,7 +475,9 @@ class Installer {
 		go()->setAuthState((new TemporaryState())->setUserId(1));
 		GO::session()->runAsRoot();
 		GO::$ignoreAclPermissions = true;
-		
+
+		go()->disableEvents();
+
 
 		$this->isValidDb();
 		go()->getCache()->flush(true, false);
@@ -483,6 +488,7 @@ class Installer {
 //		if(!empty($unavailable)) {
 //			throw new \Exception("There are unavailable modules: " . var_export($unavailable, true));
 //		}
+
 
 		$this->disableUnavailableModules();
 
@@ -554,6 +560,8 @@ class Installer {
 
 		$this->enableGarbageCollection();
 		$this->enableDiskUsage();
+
+		go()->enableEvents();
 
 		echo "Done!\n";