+++ www/modules/email/EmailClient.js
@@ -116,7 +116,8 @@ GO.email.EmailClient = Ext.extend(Ext.Panel, {
 		hidden:messagesAtTop,
 		deleteConfig : deleteConfig,
 		header:false,
-		split:true
+		split:true,
+		minWidth: dp(300)
 	});
 	this.addGridHandlers(this.leftMessagesGrid);
 
@@ -472,7 +473,8 @@ GO.email.EmailClient = Ext.extend(Ext.Panel, {
 		split: true,
 		narrowWidth: dp(400), //this will only work for panels inside another panel with layout=responsive. Not ideal but at the moment the only way I could make it work
 		width: dp(700),
-		minWidth: dp(300),
+		minWidth: dp(600),
+		narrowMinWidth: dp(300),
 		stateId: 'go-email-west',
 		items: [			
 			this.leftMessagesGrid,