+++ CHANGELOG.md
@@ -1,4 +1,22 @@
+23-05-2024: 6.8.49
+- Core: Create entity filters by users without admin privileges possible
+- Files: Fixed permissions error for admins
+- Zpushadmin: fixed error loading files
+- Core: Disable events during upgrade and install to prevent problems with modules that are not available.
+
+21-05-2024: 6.8.48
+- Email: Fixed: Feature request: icons up top in inbox (search, accounts, etc.) #947. Set minWidth for email panel
+- Notes / Comments: Use StringUtil to remove style from notes and comments as old way could remove text unexpectedly
+- Oauth: RefreshToken for Google was not obtained when using openid
+- Oauth: Possible now to use a different smtp user for IMAP accounts.
+- Oauth: disable saving to sent folder for Azure as they save sent items automatically.
+- E-mail: Client side sorting in chunks to avoid error when sorting a large search result for Microsoft Exchange server that does not support server side sort.
+- E-mail: New account option to disable saving of sent mail for Microsoft Exchange Servers because they do that on the server already.˚
+
+13-05-2024: 6.8.47
 - Email: Fixed CSS bleeding issue
+- ldap: LDAP - Synchronize users checkbox #1144
+- Core: Do not check module availability on listeners rebuild so it always rebuilds even if license fails
 
 06-05-2024: 6.8.46
 - Core: Set maximum password length to 255 to prevent brute force attacks